Implement basic tools for cosigning server
This includes code organization, logging, configuration, a testing
framework, and an initial database implementation.

The majority of code is kept in a library separate from the binary to
simplify testing. A TestBuilder is provided for Setup and Teardown when
writing tests such that resources are handled cleanly and boilerplate
minimized. A keys module is integrated with the TestBuilder for
emulating the generation of bitcoin and noise keys in place of a real
ceremony.