Update to latest revault_tx (slight behaviour change)
We pass around signatures instead of byte vectors from the database.

We use the dedicated interface to add signatures to the Spend
transaction. We now return `null` if the stored signatures are not valid
for the Spend we are presented: ie if we are polled for another
transaction spending the very same coins. Honest software would ~never
do that.